Update DependencyMap so that it actually works (#21)
The existing `Rc::into_inner` trick _never_ works, because the Replica
always holds one count of the `Rc` for itself, and returns a second
count -- so the reference count is always two. And, Rc::unwrap_or_clone`
doesn't work because `DependencyMap` can't be cloned.

This is a giant hack, as explained by comments inline. It's also
temporary, until
https://github.com/GothenburgBitFactory/taskchampion/pull/514 is
released.