Replica is not Send
The TC type `Replica` is legitimately not `Send`, because it embeds a
`Storge` implementation and that implementation may not be `Send`. In
fact, the SQLite implementation is not -- SQLite is single-threaded.

I think this is a reasonable restriction for the library for the moment.
If there is at some point a need to use a Replica from multiple threads,
then we can consider ways to lift this restriction safely.