Use PyResult for all return values to Python
In many cases, this performs a similar conversion as that supplied
automatically by pyo3, but it also allows raising ValueError for invalid
UUIDs instead of the default RuntimeError.

This also fixes the type of the `timestamp` property of
`Operation.Update`, which was missed when converting other datetimes to
use the `datetime`/`DateTime<Utc>` types.